feat: call diva-react as if it is a separate package
