+++ .eslintignore
@@ -1,0 +1,2 @@
+node_modules
+src/diva-react/node_modules
