+++ src/containers/IssueEanPage/IssueEanPage.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { withRouter } from 'react-router-dom';
 
-import IssueCredentials from '../../diva-react/containers/IssueCredentials/IssueCredentials';
+import { IssueCredentials } from '../../diva-react';
 
 const IssueEanPage = () => (
   <IssueCredentials
