+++ src/diva-react/.eslintrc
@@ -1,0 +1,25 @@
+{
+    "extends": [
+      "airbnb",
+      "plugin:jasmine/recommended",
+      "plugin:mocha/recommended"
+    ],
+    "parser": "babel-eslint",
+    "env": {
+      "es6": true,
+      "browser": true,
+      "node": true,
+      "mocha": true,
+      "jasmine": true
+    },
+    "rules": {
+      "import/no-extraneous-dependencies": ["off"],
+      "import/no-unresolved": ["off"],
+      "react/jsx-filename-extension": ["off"],
+      "react/require-default-props": ["off"]
+    },
+    "plugins": [
+      "jasmine",
+      "mocha"
+    ]
+}
