+++ src/diva-react/.gitignore
@@ -1,0 +1,25 @@
+# See https://help.github.com/ignore-files/ for more about ignoring files.
+
+# dependencies
+/node_modules
+
+# testing
+/coverage
+
+# production
+/build
+
+# misc
+.DS_Store
+.env.local
+.env.development.local
+.env.test.local
+.env.production.local
+
+# cypress
+cypress/screenshots
+cypress/videos
+
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
