+++ src/diva-react/.npmrc
@@ -1,0 +1,2 @@
+loglevel=silent
+save-exact=true
