+++ src/diva-react/.nvmrc
@@ -1,0 +1,1 @@
+8.6.0
