+++ src/diva-react/package.json
@@ -1,0 +1,48 @@
+{
+  "name": "diva-react",
+  "version": "1.0.0",
+  "description": "A frontend libary that can be used to verify IRMA attributes.",
+  "private": true,
+  "repository": {
+    "type": "git",
+    "url": "git+ssh://git@github.com/Alliander/diva-react.git"
+  },
+  "dependencies": {
+    "axios": "0.17.0",
+    "material-ui": "0.19.4",
+    "qrcode.react": "0.7.2",
+    "react": "16.3.2",
+    "react-flexbox-grid": "1.1.5",
+    "react-redux": "5.0.6",
+    "redux-saga": "0.16.0"
+  },
+  "main": "./src/index.js",
+  "scripts": {
+    "lint": "./node_modules/.bin/eslint -c .eslintrc src"
+  },
+  "keywords": [
+    "react",
+    "irma",
+    "attributes",
+    "privacy",
+    "identity",
+    "management"
+  ],
+  "contributors": [
+    "Koen van Ingen",
+    "Timen Olthof",
+    "Alliander N.V. (https://alliander.com/)"
+  ],
+  "license": "BSD-3-Clause",
+  "devDependencies": {
+    "eslint": "4.8.0",
+    "eslint-config-airbnb": "15.1.0",
+    "eslint-config-react-app": "2.0.1",
+    "eslint-plugin-flowtype": "2.39.1",
+    "eslint-plugin-import": "2.7.0",
+    "eslint-plugin-jasmine": "2.9.3",
+    "eslint-plugin-jsx-a11y": "5.1.1",
+    "eslint-plugin-mocha": "5.0.0",
+    "eslint-plugin-react": "7.4.0"
+  }
+}
