+++ src/diva-react/src/WithSimpleDivaAuthorization.js
