+++ src/diva-react/src/containers/Sign/SignCancelled.js
