+++ src/diva-react/src/index.js
@@ -1,0 +1,15 @@
+import divaSaga from './sagas/diva-saga';
+import divaReducer, { types } from './reducers/diva-reducer';
+import IssueCredentials from './containers/IssueCredentials/IssueCredentials';
+import Sign from './containers/Sign/Sign';
+import WithSimpleDivaAuthorization from './WithSimpleDivaAuthorization';
+import WithDivaAuthorization from './WithDivaAuthorization';
+
+export {
+  divaSaga,
+  divaReducer, types,
+  IssueCredentials,
+  Sign,
+  WithSimpleDivaAuthorization,
+  WithDivaAuthorization,
+};
