+++ src/diva-react/src/reducers/diva-reducer.js
