+++ src/diva-react/src/services/diva-service.js
