+++ src/sagas/index.js
@@ -2,7 +2,7 @@ import { all } from 'redux-saga/effects';
 
 import appSaga from './app-saga';
 import sessionSaga from './session-saga';
-import divaSaga from '../diva-react/sagas/diva-saga';
+import { divaSaga } from '../diva-react';
 
 const baseUrl = `${window.env.baseUrl}/api`;
